images/worker-mape-image/requirements.txt
flask
requests
psycopg2-binary
pandas